requirements.txt
langchain
langchain-community
langchain-chroma
langchain-google-vertexai
langchain-core
langchain-text-splitters
google-genai
chromadb
streamlit
pypdf
python-dotenv